third_party/WebKit/LayoutTests/fast/dom/Window/property-access-on-cached-properties-after-frame-removed.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="resources/window-property-collector.js"></script>
<script>
var childWindow;
var propertiesToVerify = [];

function insertExpectedResult(path, expected)
{
    if (path.length < 2)
        return;
    var propertyDir = path.slice(0, -1).join('.');
    var cachedPropertyDir = "cached_" + propertyDir.replace('.', '_');
    window[cachedPropertyDir] = eval("childWindow." + propertyDir);
    propertiesToVerify.push({
        'property': ["window", cachedPropertyDir, path[path.length - 1]].join("."),
        'expected': expected,
    });
}

function runTest()
{
    description("Tests access of cached DOMWindow properties after the associated frame is no longer in a web page. Test should not crash and properties should be set to sane defaults.");
    var frame = document.getElementById("frame");
    childWindow = frame.contentWindow;
    // Have expected results assume that the frame hasn't been closed (=> window.closed = false.)
    collectProperties(false);
    frame.parentNode.removeChild(frame);
    for (var i = 0; i < propertiesToVerify.length; ++i)
        shouldBe(propertiesToVerify[i].property, propertiesToVerify[i].expected);
}
</script>
</head>
<body>
<iframe id="frame" src="about:blank" onload="runTest()"></iframe>
</body>
</html>
